refactor: pr 수정요청 반영 경로수정 및 컴포넌트 분리 (#141)
