+++ src/features/chat/page/chat-list/ChatList.tsx
@@ -1,8 +1,7 @@
 import { useEffect, useRef } from "react"
 import type { ChatMessage } from "../../types/message.types"
-import MessageBubble from "./message-bubble/MessageBubble"
-import RecommendationCard from "./recommendation-card/RecommendationCard"
-import TypingBubble from "./typing-bubble/TypingBubble"
+import ChatItem from "./chat-item/ChatItem"
+
 type ChatListProps = {
   messages: ChatMessage[]
 }
@@ -23,23 +22,9 @@ const ChatList = ({ messages }: ChatListProps) => {
       ref={scrollRef}
       className="flex min-h-0 flex-1 flex-col gap-md overflow-y-auto bg-surface-default p-lg [scrollbar-gutter:stable]"
     >
-      {messages.map((message) => {
-        if (message.type === "text") {
-          return (
-            <MessageBubble
-              key={message.id}
-              role={message.role}
-              text={message.text}
-            />
-          )
-        }
-
-        if (message.type === "typing") {
-          return <TypingBubble key={message.id} />
-        }
-
-        return <RecommendationCard key={message.id} {...message.data} />
-      })}
+      {messages.map((message) => (
+        <ChatItem key={message.id} message={message} />
+      ))}
     </section>
   )
 }
