+++ src/features/photo/page/photo-tips-section/PhotoTipsSection.tsx
