+++ src/features/photo/page/photo-upload-section/mobile-camera-content/MobileCameraContent.tsx
