+++ src/features/photo/page/photo-upload-section/mobile-camera-content/camera-icon-button/CameraIconButton.tsx
