+++ src/features/photo/page/photo-upload-section/upload-preview-box/UploadPreviewBox.tsx
