+++ src/features/photo/types/mobile-photo-step.types.ts
