+++ src/features/survey/page/preference-slider/PreferenceSlider.stories.tsx
