+++ src/features/survey/page/preference-slider/PreferenceSlider.tsx
