+++ src/features/survey/page/preference-slider/preference-slider.css
