+++ src/routes/_wide.find-scent.chat.tsx
@@ -1,4 +1,4 @@
-import ScentChat from "@/features/find-scent/chat/page/ScentChat"
+import ScentChat from "@/features/chat/page/ScentChat"
 import { createFileRoute } from "@tanstack/react-router"
 
 export const Route = createFileRoute("/_wide/find-scent/chat")({
