+++ src/routes/_wide.find-scent.photo.tsx
@@ -1,4 +1,4 @@
-import ScentPhoto from "@/features/find-scent/photo/page/ScentPhoto"
+import ScentPhoto from "@/features/photo/page/ScentPhoto"
 import { createFileRoute } from "@tanstack/react-router"
 
 export const Route = createFileRoute("/_wide/find-scent/photo")({
