+++ src/routes/_wide.find-scent.survey.tsx
@@ -1,4 +1,4 @@
-import ScentSurvey from "@/features/find-scent/survey/page/ScentSurvey"
+import ScentSurvey from "@/features/survey/page/ScentSurvey"
 import { createFileRoute } from "@tanstack/react-router"
 
 export const Route = createFileRoute("/_wide/find-scent/survey")({
